Fix typo in the "Elevation to Stable status" section
One paragraph claims that interfaces can be committed to the next *minor*
release, and a following sentence claims that the accepted proposal must
be included in the next *major* release.

Change the first reference to *minor* release to *major* in order to
unify these claims about when the change to "Stable" can occur.

Fixes #5

Signed-off-by: Stephen Herbein <herbein1@llnl.gov>